+++ core/Lucy/Index/PolyLexicon.c
@@ -36,17 +36,17 @@ PolyLex_new(String *field, Vector *sub_readers) {
 
 PolyLexicon*
 PolyLex_init(PolyLexicon *self, String *field, Vector *sub_readers) {
-    uint32_t  num_sub_readers = Vec_Get_Size(sub_readers);
+    size_t    num_sub_readers = Vec_Get_Size(sub_readers);
     Vector   *seg_lexicons    = Vec_new(num_sub_readers);
 
     // Init.
     Lex_init((Lexicon*)self, field);
     PolyLexiconIVARS *const ivars = PolyLex_IVARS(self);
     ivars->term            = NULL;
-    ivars->lex_q           = SegLexQ_new(num_sub_readers);
+    ivars->lex_q           = SegLexQ_new((uint32_t)num_sub_readers);
 
     // Derive.
-    for (uint32_t i = 0; i < num_sub_readers; i++) {
+    for (size_t i = 0; i < num_sub_readers; i++) {
         LexiconReader *lex_reader = (LexiconReader*)Vec_Fetch(sub_readers, i);
         if (lex_reader && CERTIFY(lex_reader, LEXICONREADER)) {
             Lexicon *seg_lexicon = LexReader_Lexicon(lex_reader, field, NULL);
@@ -192,7 +192,7 @@ PolyLex_Get_Term_IMP(PolyLexicon *self) {
 uint32_t
 PolyLex_Get_Num_Seg_Lexicons_IMP(PolyLexicon *self) {
     PolyLexiconIVARS *const ivars = PolyLex_IVARS(self);
-    return Vec_Get_Size(ivars->seg_lexicons);
+    return (uint32_t)Vec_Get_Size(ivars->seg_lexicons);
 }
 
 SegLexQueue*
